refactor(core): converge Python responsibility boundaries
