+++ .github/workflows/python-structure.yml
@@ -1,0 +1,54 @@
+# SPDX-License-Identifier: Apache-2.0
+
+name: Python Structure
+
+on:
+  pull_request:
+    branches:
+      - dev/v*/v*
+      - alpha/v*/v*
+      - release/v*/v*
+    paths:
+      - "framework/core/src/python/**"
+      - "framework/core/tests/python/**"
+      - "framework/maintainability/abstraction-integrity*.json"
+      - "framework/maintainability/python-structure-negative-fixtures.json"
+      - "scripts/check-code-complexity*.mjs"
+      - "scripts/check-python-structure*.py"
+      - ".github/workflows/python-structure.yml"
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+concurrency:
+  group: python-structure-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  exact-root-ratchet:
+    name: Exact-root ratchet (${{ matrix.os }})
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-24.04, macos-14, windows-2022]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Check out exact candidate with baseline history
+        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+        with:
+          fetch-depth: 0
+          persist-credentials: false
+      - uses: actions/setup-node@49933ea5288caeca8642d1e84afbd3f7d6820020 # v4.4.0
+        with:
+          node-version: 24
+      - uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065 # v5.6.0
+        with:
+          python-version: "3.13"
+      - name: Verify Python structure gate (POSIX)
+        if: runner.os != 'Windows'
+        run: ./shifu check:python-structure
+      - name: Verify Python structure gate (Windows)
+        if: runner.os == 'Windows'
+        shell: pwsh
+        run: .\shifu.cmd check:python-structure
