+++ framework/core/src/python/kungfu/assignment_lifecycle/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Owned contracts shared by native Assignment lifecycle services."""
