+++ framework/release/publication-surfaces.json
@@ -283,6 +283,13 @@
           "sourceRoot": "sha256:1700c2c52b5ff4b09016013b2cf643fc570560a31c805434153d524a83e2f073"
         },
         {
+          "path": ".github/workflows/python-structure.yml",
+          "classification": "non-publication",
+          "owner": "core-python",
+          "surfaceIds": [],
+          "sourceRoot": "sha256:e85ad35d75c680bd5890462ff2f3d8162751cff3a2fb1f579578045a841d621a"
+        },
+        {
           "path": ".github/workflows/release-new-version.yml",
           "classification": "publication",
           "owner": "kungfu-release",
@@ -821,5 +828,5 @@
       }
     }
   ],
-  "registryRoot": "sha256:c42ee2bce33ecddb01dc141ec5ea54fd63b49d19a0378b31cc00eb1c5f0d3761"
+  "registryRoot": "sha256:eb2138871147f4ab71da8f2b3d8bf0c97b32c26b03f1131764af5368387e0d04"
 }
