+++ package.json
@@ -181,6 +181,8 @@
     "check:work-agent-history-continuity": "node scripts/require-shifu.mjs check:work-agent-history-continuity && node --test scripts/check-work-agent-history-continuity.test.mjs && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} KUNGFU_NATIVE_PATH=${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_run_product_path.py::test_agent_activity_history_projection_keeps_process_success_outside_work framework/core/tests/python/test_agent_work_profile_native.py::test_native_profile_authority_bundle_restores_clean_home_exactly framework/core/tests/python/test_agent_work_profile_native.py::test_native_profile_authority_import_preflights_all_operations_before_writing && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} KUNGFU_NATIVE_PATH=${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_agent_work_state_contract.py::test_kfd7_legacy_role_roots_remain_readable_without_reinterpretation",
     "check:project-cut-dogfood-history": "node scripts/require-shifu.mjs check:project-cut-dogfood-history && node --test scripts/check-project-cut-dogfood-history.test.mjs scripts/check-data-protection-contract.test.mjs scripts/check-exit-bundle-contract.test.mjs && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} KUNGFU_NATIVE_PATH=${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_exit_bundle.py::test_project_cut_history_member_survives_source_removal framework/core/tests/python/test_exit_bundle.py::test_project_cut_history_member_rejects_thin_and_diverged_destinations framework/core/tests/python/test_exit_bundle.py::test_dogfood_profile_facts_roundtrip_through_native_exit_members",
     "check:product-release-cut-portability": "node scripts/require-shifu.mjs check:product-release-cut-portability && node --test scripts/check-product-release-cut-portability.test.mjs scripts/check-data-protection-contract.test.mjs scripts/check-exit-bundle-contract.test.mjs && PYTHONPATH=$PWD/framework/core/src/python:$PWD/framework/core/build/Release KUNGFU_NATIVE_PATH=$PWD/framework/core/build/Release $PWD/framework/core/.venv/bin/python -m pytest -q framework/core/tests/python/test_product_release_history.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} KUNGFU_NATIVE_PATH=${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_release_cut.py::test_shifu_local_successor_is_explicitly_publication_ineligible framework/core/tests/python/test_release_cut.py::test_shifu_local_bootstrap_is_explicit_and_cannot_target_public_cut",
+    "check:python-structure": "node scripts/require-shifu.mjs check:python-structure && node scripts/check-code-complexity.mjs && node --test scripts/check-code-complexity.test.mjs",
+    "maintainability:python-structure": "node scripts/require-shifu.mjs maintainability:python-structure && node scripts/check-code-complexity.mjs",
     "check:exit-bundle-contract": "node scripts/require-shifu.mjs check:exit-bundle-contract && node --test scripts/check-exit-bundle-contract.test.mjs",
     "check:exit-verifier-package": "node scripts/require-shifu.mjs check:exit-verifier-package && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen python scripts/check-exit-verifier-package.py --json",
     "check:exit-clean-runtime": "node scripts/require-shifu.mjs check:exit-clean-runtime && python3 scripts/check-exit-clean-runtime.py --json",
