+++ scripts/readonly-agent-bootstrap.test.mjs
@@ -218,6 +218,7 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     '.github/workflows/gate-measurement.yml',
     '.github/workflows/build.yml',
     '.github/workflows/publish-layer-artifacts.yml',
+    '.github/workflows/python-structure.yml',
     '.github/workflows/release-new-version.yml',
     '.github/workflows/stable-candidate-patrol.yml',
     '.github/actionlint.yaml',
@@ -296,6 +297,9 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     'scripts/kungfu-invariant.mjs',
     'scripts/code-complexity-budget.mjs',
     'scripts/code-complexity-budget.test.mjs',
+    'scripts/check-code-complexity.mjs',
+    'scripts/check-code-complexity.test.mjs',
+    'scripts/check-python-structure.py',
     'scripts/check-readonly-source-routes.mjs',
     'scripts/check-readonly-source-routes.test.mjs',
     'scripts/check-alpha-attention-operations.mjs',
@@ -388,6 +392,10 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     'framework/maintainability/waivers/retired/hub-cli-linux-arm64-workflow-authority.v1.json',
     'framework/maintainability/waivers/retired/libnode-linux-arm64-lockfile.v1.json',
     'framework/maintainability/complexity-governance.mjs',
+    'framework/maintainability/abstraction-integrity.manifest.json',
+    'framework/maintainability/abstraction-integrity.baseline.json',
+    'framework/maintainability/abstraction-integrity-report.json',
+    'framework/maintainability/python-structure-negative-fixtures.json',
     'framework/maintainability/readonly-source-routes.json',
     'framework/maintainability/semantic-amplification.manifest.json',
     'framework/maintainability/semantic-amplification-report.json',
